backend/requirements.txt
Flask>=2.0
Werkzeug>=2.0
Flask-Limiter>=2.0
Flask-Mail>=0.9
python-dotenv>=0.19